requirements_streamlit.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pandas>=2.0.0
pyarrow>=12.0.0
google-generativeai>=0.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
evaluate>=0.4.0
bert-score>=0.3.13
sacrebleu>=2.3.1
nltk>=3.8.1
sacremoses>=0.0.53

